+++ src/server/game/Handlers/MailHandler.cpp
@@ -163,7 +163,7 @@ void WorldSession::HandleSendMail(WorldPacket& recvData)
         return;
     }
 
-    uint32 cost = items_count ? 30 * items_count : 30;  // price hardcoded in client
+    uint64 cost = items_count ? 30 * items_count : 30;  // price hardcoded in client
 
     uint64 reqmoney = cost + money;
